requirements.txt
Django>=3.2.3,<3.3
djangorestframework>=3.14.0,<3.15
Pillow>=9.3.0, <9.4
psycopg2>=2.8.6,<2.9
requests>=2.28.1, <2.29